Compute all pairwise hamming distances between matrix rows.

Returns a distance matrix of all meanings. The meanings argument should be

hammingdists(meanings)

Arguments

meanings
a matrix with the different dimensions encoded along columns, and all combinations of meanings specified along rows. The data type of the cells does not matter since distance is simply based on equality - in fact specifying a meaning component as NA allows you to ignore that dimension for the given row/meaning combinations (see examples). Vectors are treated as matrices with a single row, so the resulting distances can only be 0 or 1.

Value

a distance matrix of type dist with n*(n-1)/2 rows/columns, where n is the number of rows in meanings.

See also

dist

Examples

# example #1: a 2x2 design using strings (the character redundantly
# specifies the dimension for clarity)
hammingdists(matrix(c("a1", "b1", "a1", "b2", "a2", "b1", "a2", "b2"), ncol=2, byrow=TRUE))

#>   1 2 3
#> 2 1    
#> 3 1 2  
#> 4 2 1 1

# example #2: a 2x3 design using integers for encoding
hammingdists(matrix(c(0,0,0,1,0,2,1,0,1,1,1,2), ncol=2, byrow=TRUE))

#>   1 2 3 4 5
#> 2 1        
#> 3 1 1      
#> 4 1 2 2    
#> 5 2 1 2 1  
#> 6 2 2 1 1 1

# example #3: a 2x2x2 design using factors (ncol is always the number of dimensions)
hammingdists(matrix(c(0,0,0,0,0,1,0,1,0,0,1,1,1,0,0,1,0,1,1,1,0,1,1,1), ncol=3, byrow=TRUE))

#>   1 2 3 4 5 6 7
#> 2 1            
#> 3 1 2          
#> 4 2 1 1        
#> 5 1 2 2 3      
#> 6 2 1 3 2 1    
#> 7 2 3 1 2 1 2  
#> 8 3 2 2 1 2 1 1

# example #4: if some meaning dimension is not relevant for some
# combinations of meanings (e.g. optional arguments), specifying them as NA
# in the matrix will make them not be counted towards the hamming distance!
# in this example the value of the second dimension does not matter (and
# does not count towards the distance) when the the first dimension has
# value '1'
hammingdists(matrix(c(0,0,0,1,1,NA), ncol=2, byrow=TRUE))

#>   1 2
#> 2 1  
#> 3 1 1