Find most predictive segments for every meaning feature.

For every meaning find the segment(s) that maximise p(m|s) * p(s|m)

substring.predictability(strings, meanings)

meaning.predictability(strings, meanings)

mostpredictivesegments(strings, meanings, verbose = FALSE)

Arguments

strings
a list or vector of character sequences

meanings
a matrix or data frame with as many rows as there are strings (see Details)

Value

a matrix with one row for every meaning feature

Functions

substring.predictability: Calculate predictability for all substrings given meaning features.

Calculate P(s_ij | f) for every s_ij in S and f in M.

Whatever the actual types of the columns, they will be treated as categorical factors.

meaning.predictability: Calculate P(f | s_ij) for every f in M and s_ij in S.

Examples

substring.predictability(c("asd", "basd", "basdf"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>         a   b
#> a     1.0 1.0
#> as    1.0 1.0
#> asd   1.0 1.0
#> s     1.0 1.0
#> sd    1.0 1.0
#> d     1.0 1.0
#> b     0.5 1.0
#> ba    0.5 1.0
#> bas   0.5 1.0
#> basd  0.5 1.0
#> basdf 0.5 0.5
#> asdf  0.5 0.5
#> sdf   0.5 0.5
#> df    0.5 0.5
#> f     0.5 0.5

meaning.predictability(c("asd", "basd", "basdf"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>           a        as       asd         s        sd         d   b  ba bas basd
#> a 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.5 0.5 0.5  0.5
#> b 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 0.6666667 1.0 1.0 1.0  1.0
#>   basdf asdf sdf df f
#> a     1    1   1  1 1
#> b     1    1   1  1 1

mostpredictivesegments(c("as", "bas", "basf"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>   meaning         p   segments
#> a       a 0.6666667   a, as, s
#> b       b 1.0000000 b, ba, bas

mostpredictivesegments(c("ab", "ac", "bc", "bd"),
  cbind(a=c(T, T, F, F), b=c(F, F, T, T)))

#>   meaning         p segments
#> a       a 1.0000000        a
#> b       b 0.6666667        b