Perform the Page test of monotonicity.

The Page test tests the given matrix for monotonically increasing ranks across k linearly ordered conditions (along columns) based on N replications (along rows). To test for monotonically decreasing trends you can either reverse the order of columns, or simply invert the ranks (i.e. by prepending the dataset with a -).

page.test(data, ...)

Arguments

data
a matrix or dataframe with the different conditions along columns and the replications along rows. Conversion of the data to ranks is taken care of internally (see page.L)

...
extra parameters which are passed on to page.L

Value

a list of class pagetest (and htest) containing the following fields:

L

Value of the L statistic for the given data set

k

Number of conditions (columns of the data set)

N

Number of replications (rows of the data set)

p.value

Significance level

p.type

Whether the computed p value is "exact" or "approx"

Details

Exact p values are computed for k up to 22, using the pre-computed null distributions from the pspearman package. For larger k, p values are computed based on a Normal distribution approximation.

Examples

page.test(t(replicate(4, sample(4))))

#> Ranking used:
#>     [,1] [,2] [,3] [,4]
#> N=1    4    2    1    3
#> N=2    1    3    4    2
#> N=3    2    3    4    1
#> N=4    1    2    4    3

#> 
#> 	Page test of monotonicity
#> 
#> data:  
#> = 103, p-value = 0.3375
#> alternative hypothesis: 
#> 

page.test(t(replicate(4, sample(10))))

#> Ranking used:
#>     [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> N=1    6    1    4    7    9   10    3    2    5     8
#> N=2   10    7    8    1    2    5    3    6    9     4
#> N=3    2   10    3    5    4    7    6    8    1     9
#> N=4    1    6    8    2    3    7   10    5    4     9

#> 
#> 	Page test of monotonicity
#> 
#> data:  
#> = 1253, p-value = 0.2226
#> alternative hypothesis: 
#>