Spike's measure of additive compositionality.

Implementation of the Spike-Montague information-theoretic measure of additive compositionality (Spike 2016), which finds the most predictive association between substrings and categorical meaning features. Additive means that it does not take the ordering of elements into account, i.e. GREEN DOG = GREEN + DOG = DOG + GREEN.

sm.compositionality(strings, meanings, groups = NULL)

sm.segmentation(strings, meanings)

Arguments

strings
a list or vector of character sequences

meanings
a matrix or data frame with as many rows as there are strings (see below)

groups
a list or vector with as many items as strings, used to split strings and meanings into data sets for which compositionality measures are computed separately.

Details

The measure really captures the degree to which a synonymy-free signalling system exists at the level of semantic features, rather than looking for complex meanings per se. The resulting segmentations may therefore overlap.

The segmentation algorithm scans through all sub-strings found in strings to find the most bijective mapping onto all the meaning features present in meanings, i.e. the pairings of sub-strings and meaning features that are most predictive of each other. Mathematically, for every meaning feature \(f\in M\), it tries to find the sub-string \(s_{ij}\) from the set of strings \(S\) that maximises $$comp(f,S) = \max_{s_{ij}\in S} P(f|s_{ij}) \cdot P(s_{ij}|f)$$.

Functions

sm.compositionality: Calculates the mean predictability of all meaning features, \(C = \frac{1}{|M|} \sum_{f\in M} comp(m,S)\). Returns a vector of three elements: N, the number of signal-meaning pairings on which the computation was based, comp, the compositionality measure, and M, the number of distinct meaning features over which the measure was computed. (When groups is not NULL, returns a matrix with the same elements along columns, with one row for every group.)

sm.segmentation: Finds the most predictive segment(s) for every meaning feature, i.e. the substrings \(s_{ij}\) that maximise \(P(m|s_{ij}) \cdot P(s_{ij}|m)\). Returns a matrix with one row for every meaning feature, in descending order of their predictability from (and to) their corresponding string segments.

Meaning data format

The meanings can be a matrix or data frame in one of two formats. If it is a matrix of logicals (TRUE/FALSE values), then the columns are assumed to refer to meaning features, with individual cells indicating whether the meaning feature is present or absent in the message indicated by that row (see binaryfeaturematrix for an explanation). If meanings is a data frame or matrix of any other type, it is assumed that the columns specify different meaning dimensions, with the cell values showing the levels with which the different dimensions can be realised. This dimension-based representation is automatically converted to a feature-based one using binaryfeaturematrix. As a consequence, whatever the actual types of the columns in the meaning matrix, they will be treated as categorical factors in order to represent them as atomic meaning features.

References

Spike, M. (2016). Minimal requirements for the cultural evolution of language. PhD thesis, The University of Edinburgh.

See also

binaryfeaturematrix

Examples

# perfect communication system
sm.compositionality(c("a", "b", "ab"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>    N comp    M 
#>    3    1    2 

sm.segmentation(c("a", "b", "ab"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>   N p segments tiedsegments
#> a 2 1        a            1
#> b 2 1        b            1

sm.compositionality(c("as", "bas", "basf"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>         N      comp         M 
#> 3.0000000 0.8333333 2.0000000 

sm.segmentation(c("as", "bas", "basf"),
  cbind(a=c(T, F, T), b=c(F, T, T)))

#>   N         p   segments tiedsegments
#> b 2 1.0000000 bas, ba, b            3
#> a 2 0.6666667   as, a, s            3

# the function also accepts meaning-dimension based matrix definitions:
enumerate.meaningcombinations(c(animal=2, colour=2))

#>      colour animal
#> [1,]      1      3
#> [2,]      1      4
#> [3,]      2      3
#> [4,]      2      4

sm.segmentation(c("greendog", "bluedog", "greencat", "bluecat"),
  enumerate.meaningcombinations(c(animal=2, colour=2)))

#>          N p     segments tiedsegments
#> colour=1 2 1 dog, do,....            5
#> colour=2 2 1 cat, ca,....            6
#> animal=3 2 1 green, g....           12
#> animal=4 2 1 blue, bl....            9