Spike's measure of additive compositionality.

Implementation of the Spike-Montague information-theoretic measure of additive compositionality (Spike 2016), which finds the most predictive association between substrings and categorical meaning features. Additive means that it does not take the ordering of elements into account, i.e. GREEN DOG = GREEN + DOG = DOG + GREEN.

sm.compositionality(x, y, groups = NULL, strict = FALSE)

sm.segmentation(x, y, strict = FALSE)

Arguments

x
a list or vector of character sequences specifying the signals to be analysed. Alternatively, x can also be a formula of the format s ~ m1 + m2 + ..., where s and m* specify the column names of the signals and meaning features found in the data frame that's passed as the second argument.

y
a matrix or data frame with as many rows as there are signals, indicating the presence/value of the different meaning dimensions along columns (see section Meaning data format). If x is a formula, the second argument can contain any number of columns, only the ones specified in the formula will be considered.

groups
a list or vector with as many items as strings, used to split strings and meanings into data sets for which compositionality measures are computed separately.

strict
logical: if TRUE, perform additional filtering of candidate segments. In particular, it removes combinations of segments (across meanings) which overlap in at least one of the strings where they co-occur. For convenience, it also removes segments which are shorter substrings of longer candidates (for the same meaning feature).

Value

sm.compositionality calculates the mean predictability of all meaning features and their most predictably co-occurring strings. Returns a vector of three elements: N is the number of signal-meaning pairings on which the computation was based, comp the mean mutual predictability, and M the number of distinct meaning features over which this average was computed. (When groups is not NULL, returns a matrix with the same elements along columns, with one row for every group.)

sm.segmentation provides detailed information about the most predictably co-occurring segments for every meaning feature. It returns a data frame with one row for every meaning feature, in descending order of their predictability from (and to) their corresponding string segments. The data frame has the following columns:

N

The number of signal-meaning pairings in which this meaning feature was attested.

mp

The highest mutual predictability between this meaning feature and one (or more) segments that was found.

p

Significance levels of the mutual predictability for the given segment(s). The calculation depends on the frequency of the meaning feature as well as the overall frequency of the segment across all signals and indicates the (null) probability of the given co-occurrence count of the two according to the hypergeometric distribution. If multiple segments are tied for their mutual predictability, the candidate list is filtered to include only those with the highest significance levels (i.e. the lowest p values).

ties

The number of substrings found in strings which have this same level of mutual predictability with the meaning feature.

segments

For strict = FALSE: a list containing the ties substrings in descending order of their length (the ordering is for convenience only and not inherently meaningful). When strict = TRUE, the lists of segments for each meaning feature are all of the same length, with a meaningful relationship of the order of segments across the different rows: every set of segments which are found in the same position for each of the different meaning features constitute a valid segmentation where the segments occurrences in the actual signals do not overlap.

Details

The measure really captures the degree to which a synonymy-free signalling system exists at the level of semantic features, rather than looking for complex meanings per se. The resulting segmentations may therefore overlap, which means holistic signalling systems can currently still get relative high compositionality scores. A modification to account for this is under development.

The segmentation algorithm scans through all sub-strings found in strings to find the most bijective mapping onto all the meaning features present in meanings, i.e. the pairings of sub-strings and meaning features that are most predictive of each other. Mathematically, for every meaning feature \(f\in M\), it tries to find the sub-string \(s_{ij}\) from the set of strings \(S\) that maximises $$comp(f,S) = \max_{s_{ij}\in S}\ P(f|s_{ij}) \cdot P(s_{ij}|f)$$.

Based on these mappings onto individual meaning features, sm.compositionality then computes the average mutual predictability over all meaning features \(M\), $$comp(M,S) = \frac{1}{|M|} \sum_{f\in M} comp(f,S),$$ as a measure of the overall compositionality of the signalling system.

Meaning data format

The meanings can be a matrix or data frame in one of two formats. If it is a matrix of logicals (TRUE/FALSE values), then the columns are assumed to refer to meaning features, with individual cells indicating whether the meaning feature is present or absent in the message indicated by that row (see binaryfeaturematrix for an explanation). If meanings is a data frame or matrix of any other type, it is assumed that the columns specify different meaning dimensions, with the cell values showing the levels with which the different dimensions can be realised. This dimension-based representation is automatically converted to a feature-based one using binaryfeaturematrix. As a consequence, whatever the actual types of the columns in the meaning matrix, they will be treated as categorical factors in order to represent them as atomic meaning features.

References

Spike, M. (2016). Minimal requirements for the cultural evolution of language. PhD thesis, The University of Edinburgh.

See also

binaryfeaturematrix, ssm.compositionality

Examples

# perfect communication system
sm.compositionality(c("a", "b", "ab"),
  cbind(a=c(TRUE, FALSE, TRUE), b=c(FALSE, TRUE, TRUE)))

#>   N comp M
#> 1 3    1 2

sm.segmentation(c("a", "b", "ab"),
  cbind(a=c(TRUE, FALSE, TRUE), b=c(FALSE, TRUE, TRUE)))

#>   N mp     p ties segments
#> a 2  1 0.333    1        a
#> b 2  1 0.333    1        b
#> 
#> Mean feature-wise mutual predictability, weighted by feature frequency: 1 

# not quite perfect communication system
sm.compositionality(c("as", "bas", "basf"),
  cbind(a=c(TRUE, FALSE, TRUE), b=c(FALSE, TRUE, TRUE)))

#>   N      comp M
#> 1 3 0.8333333 2

sm.segmentation(c("as", "bas", "basf"),
  cbind(a=c(TRUE, FALSE, TRUE), b=c(FALSE, TRUE, TRUE)))

#>   N    mp     p ties   segments
#> b 2 1.000 0.333    3 bas, ba, b
#> a 2 0.667 1.000    3   as, a, s
#> 
#> Mean feature-wise mutual predictability, weighted by feature frequency: 0.833 

# force candidate segments to be non-overlapping via the 'strict' option
sm.segmentation(c("as", "bas", "basf"),
  cbind(a=c(TRUE, FALSE, TRUE), b=c(FALSE, TRUE, TRUE)), strict=TRUE)

#> Applying strict selection, checking 9 segment combinations for overlap

#>   N    mp     p ties segments
#> b 2 1.000 0.333    2    s, ba
#> a 2 0.667 1.000    2    as, b
#> 
#> Mean feature-wise mutual predictability, weighted by feature frequency: 0.833 

# the function also accepts meaning-dimension based matrix definitions:
print(twobytwoanimals <- enumerate.meaningcombinations(c(animal=2, colour=2)))

#>      colour animal
#> [1,]      1      3
#> [2,]      1      4
#> [3,]      2      3
#> [4,]      2      4

# note how there are many more candidate segments than just the full length
# ones. given limited data, it is expected that shorter substrings will be
# just as predictable as the full segments that contain them.
sm.segmentation(c("greendog", "greencat", "bluedog", "bluecat"), twobytwoanimals)

#>          N mp     p ties     segments
#> colour=1 2  1 0.167   12 green, g....
#> colour=2 2  1 0.167    9 blue, bl....
#> animal=3 2  1 0.167    5 dog, do,....
#> animal=4 2  1 0.167    6 cat, ca,....
#> 
#> Mean feature-wise mutual predictability, weighted by feature frequency: 1 

# since there is no overlap in the constituent characters of the identified
# 'morphemes', they are all tied in their mutual predictiveness with the
# (shorter) substrings they contain

# to reduce the pool of candidate segments to those which are
# non-overlapping and of maximal length, again use the 'strict=TRUE' option:

sm.segmentation(c("greendog", "greencat", "bluedog", "bluecat"), twobytwoanimals,
  strict=TRUE)

#> Applying strict selection, checking 3240 segment combinations for overlap

#>          N mp     p ties segments
#> colour=1 2  1 0.167    1    green
#> colour=2 2  1 0.167    1     blue
#> animal=3 2  1 0.167    1      dog
#> animal=4 2  1 0.167    1      cat
#> 
#> Mean feature-wise mutual predictability, weighted by feature frequency: 1